`render_priority`: f64 -> f32

+++ mitmproxy-contentviews/src/hex_dump.rs
@@ -24,7 +24,7 @@ impl Prettify for HexDump {
         ))
     }
 
-    fn render_priority(&self, data: &[u8], _metadata: &dyn Metadata) -> f64 {
+    fn render_priority(&self, data: &[u8], _metadata: &dyn Metadata) -> f32 {
         if is_binary(data) {
             0.5
         } else {
